# api/src/main/resources/api-definition.yaml
openapi: 3.0.3
info:
  contact:
    name: Sandro Andreasik
  description: Sprint Planning Poker for Scrum teams
  title: EFI Poker API
  version: 1.0.0
servers:
- description: API base path
  url: /api/v1
...
paths:
...
  /projects:
    post:
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectAdminResponse"
          description: Project created
      summary: Create a new project
      tags:
      - projects
  /projects/{slug}:
    delete:
      operationId: deleteProject
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "204":
          description: Project deleted
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Delete project and all associated data (admin)
      tags:
      - projects
    get:
      operationId: getProjectBySlug
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
          description: Project info
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Get project info
      tags:
      - projects
    patch:
      operationId: updateProject
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectAdminResponse"
          description: Updated project
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Update project (admin)
      tags:
      - projects
  /projects/{slug}/admin:
    get:
      operationId: getProjectAdmin
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      - in: header
        name: X-Admin-Code
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectAdminResponse"
          description: Project admin info
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Access denied
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Resource not found
      summary: Get project with admin data
      tags:
      - projects
...
  /auth/me/projects:
    get:
      description: Returns all projects where the authenticated user is the owner.
      operationId: getMyProjects
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ProjectAdminResponse"
                type: array
          description: List of owned projects
      security:
      - bearerAuth: []
      summary: Get projects owned by current user
      tags:
      - projects
  /auth/me/participated-projects:
    get:
      description: Returns projects the authenticated user has joined as a participant.
      operationId: getMyParticipatedProjects
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ProjectResponse"
                type: array
          description: List of participated projects
      security:
      - bearerAuth: []
      summary: Get projects where current user is a participant
      tags:
      - projects
...
components:
...
  schemas:
...
    CreateProjectRequest:
      example:
        name: name
      properties:
        name:
          maxLength: 255
          minLength: 1
          type: string
      required:
      - name
      type: object
    ProjectAdminResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        adminCode: adminCode
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        slug: slug
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        slug:
          type: string
        adminCode:
          type: string
        createdAt:
          format: date-time
          type: string
      type: object
    ProjectResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        slug: slug
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        slug:
          type: string
        createdAt:
          format: date-time
          type: string
      type: object
    ProblemDetail:
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      type: object
    UpdateProjectRequest:
      example:
        name: name
      properties:
        name:
          maxLength: 255
          minLength: 1
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http